chore: rename BleDevicesScanning
