+++ apps/ledger-live-mobile/src/screens/BleDevicePairingFlow/BleDevicesScanning.tsx
